requirements.txt
flask==3.0.3
flask-cors==5.0.0
flask-jwt-extended==4.6.0
flask-sqlalchemy==3.1.1
flask-migrate==4.0.7
python-dotenv==1.0.1
bcrypt==4.2.0
requests==2.32.3
gunicorn==22.0.0
